https://hkpark130.p-e.kr:8080/issues/24

+++ src/main/java/kr/pe/hkpark130/springblog/domain/comment/CommentsRepository.java
@@ -1,6 +1,5 @@
 package kr.pe.hkpark130.springblog.domain.comment;
 
-import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -11,6 +10,6 @@ import java.util.List;
 public interface CommentsRepository extends JpaRepository<Comments, Long> {
     @Transactional
     @Query(value = "SELECT c.* FROM comments c WHERE c.post_id = :post_id", nativeQuery = true)
-    List<Comments> findAll(@Param("post_id") Integer post_id);
+    List<Comments> findAll(@Param("post_id") Long post_id);
 
 }
